Extract ComposeSceneAccessible
